+++ lib/utils/Utils.dart
@@ -262,3 +262,10 @@ class NullableValue<T> {
 
   factory  NullableValue.empty() => NullableValue(null);
 }
+
+enum ContentActivity { reload, refresh, extend }
+extension ContentActivityImpl on ContentActivity {
+  bool get loading => (this == ContentActivity.reload) || (this == ContentActivity.refresh);
+  bool get extending => (this == ContentActivity.extend);
+}
+